Hack in support for pilot vs. passenger
Not the best implementation, but Operational Transformations are hard. I
decided not to re-implement yet another live collaboration tool and
instead focus on the fact that only one person has the keyboard anyway.
The UI needs to make it more obvious that it's read only when you're not
a pilot. I'm seriously reconsidering doing the frontend in React
actually.